Remove ArchUser and use django-registration
The ArchUser User profile added complexities for little benefit (IRC username
and inactivity). It was removed in favor of aurprofile for AUR user profiles.
This new app provides only provides a profile page and a form for updating
user details.

Handling of the registration process has been out-sourced to the third-party
django-registration app, which does more than the previous system did. This
obviously adds a dependency for that app.

Signed-off-by: Sebastian Nowicki <sebnow@gmail.com>