Change categories to tags
This allows for a much more flexible way of categorizing packages. If
pacman/makepkg ever support categories, it will be much easier to integrate
that data.

This uses the django-tagging third party app, which means an additional
dependency. The app does not support non-integer based primary keys for
taggable models, thus Package's primary key was changed to the standard "id"
and name was changed to be unique. A slug was also added to keep the
object_detail generic view working.

Note: This is a schema backwards-incompatible change.