Re-organize settings file
Instead of tracking changes in settings.py.proto and merging with a
local settings.py, the settings.py is tracked and local settings are
imported from settings_local.py. The imported settings can override the
settings defined in settings.py, so it's still flexible.

Better efaults have also been defined, for a development environment.
This makes it faster and easier to start developing.

Signed-off-by: Sebastian Nowicki <sebnow@gmail.com>