Add voting for packages.
A Vote object records the fact that a given user voted for a certain
package at a known time. Voting and unvoting are implemented. Also, the
details page for a package shows the number of votes that it has
received.

This patch incorporates the following changes by Ismael Carnales:
1. Use get_package_or_404 in unvote, from views.py.  Trying to remove a
vote for a non-existent package should produce a 404 HTTP response.
2. The timestamp in the Vote class should be named "added" for
consistency, rather than "when", which was the name used in the
original patch.

This patch inforporates the following changes by Sebastian Nowicki:
1. Use "default=datetime.now" as opposed to "auto_now_add=True" for
added field.
2. Use a slug instead of the object_id in urls.
3. Register the Vote model with the django admin interface.

Signed-off-by: Christopher Brannon <cmbrannon79@gmail.com>
Signed-off-by: Sebastian Nowicki <sebnow@gmail.com>