Fix login error message
The issue was that no error appeared after trying to log in with an
invalid username or password, just the prompt was got back. This is
because the "has_errors" form attribute was renamed to "errors" in
Django 1.0.

Changing form.has_errors to form.error in the login template solved the
problem

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>
Signed-off-by: Sebastian Nowicki <sebnow@gmail.com>